Fix cache element invalidation bug.
Although we only want to cache compatible full elements, invalidation
needs to occur even when an element is given that has a superextent of
the vertex to be invalidate. This typically happens when invalidation of
an ancestor compute X propagates to a descendant compute Y where Y's
scope includes subextents of the extents of X's scope.